separated mat bleed into its own file, changed where mat write method is overridden from inside FModel to its initialization calls
