+++ fast64_internal/sm64/sm64_geolayout_classes.py
@@ -17,7 +17,7 @@ from ..utility import (
     radians_to_s16,
     geoNodeRotateOrder,
 )
-from ..f3d.f3d_gbi import BleedGraphics
+from ..f3d.f3d_bleed import BleedGraphics
 
 from .sm64_geolayout_constants import (
     nodeGroupCmds,
@@ -463,7 +463,9 @@ class GeoLayoutBleed(BleedGraphics):
                 lastMat = last_materials.get(base_node.drawLayer, None)
                 default_render_mode = fModel.getRenderMode(base_node.drawLayer)
                 lastMat = self.bleed_fmesh(fModel.f3d, fMesh, lastMat, cmd_list, default_render_mode)
-                last_materials[base_node.drawLayer] = lastMat
+                # if the mesh has culling, it can be culled, and create invalid combinations of f3d to represent the current full DL
+                if not fMesh.cullVertexList:
+                    last_materials[base_node.drawLayer] = lastMat
             # don't carry over lastmat if it is a switch node or geo asm node
             if type(base_node) in [SwitchNode, FunctionNode, JumpNode]:
                 last_materials = dict()
